添加生命值系统，人物跳跃功能，初步添加伤害系统

+++ Assets/Scripts/Manager/UI/PlayerUIHudManager.cs
@@ -18,7 +18,7 @@ public class PlayerUIHudManager : MonoBehaviour
         LayoutRebuilder.ForceRebuildLayoutImmediate(layoutGroup.GetComponent<RectTransform>());
     }
     #region ÑªÁ¿
-    public void SetNewHealthValue(int oldValue, int newValue)
+    public void SetNewHealthValue(float oldValue, float newValue)
     {
         healthBar.SetStat(newValue);
     }
